requirements.txt
unstructured[pdf]
google-api-python-client
pymupdf